move install command into 'create-startupjs' package
