+++ packages/cli/commands/index.js
@@ -4,4 +4,4 @@ export * as task from './pm/task.js'
 export * as postinstall from './postinstall.js'
 export * as startProduction from './start-production.js'
 export * as build from './build.js'
-export * as install from './install/index.js'
+export * as install from 'create-startupjs/install'
