+++ packages/create-startupjs/create/templates/blank/.env
