+++ packages/create-startupjs/create/templates/blank/.eslintrc.json
