+++ packages/create-startupjs/create/templates/blank/App.js
