+++ packages/create-startupjs/create/templates/blank/babel.config.js
