+++ packages/create-startupjs/create/templates/blank/components/TestComponent/index.js
