+++ packages/create-startupjs/create/templates/blank/components/index.js
