+++ packages/create-startupjs/create/templates/blank/config.json
