+++ packages/create-startupjs/create/templates/blank/index.js
