+++ packages/create-startupjs/create/templates/blank/model/TestThingModel.js
