+++ packages/create-startupjs/create/templates/blank/model/index.js
