+++ packages/create-startupjs/create/templates/blank/server.js
