+++ packages/create-startupjs/create/templates/blank/server/api/index.js
