+++ packages/create-startupjs/create/templates/blank/server/api/testThing.js
