+++ packages/create-startupjs/create/templates/blank/server/index.js
