+++ packages/create-startupjs/create/templates/blank/styles/index.styl
