+++ packages/create-startupjs/create/templates/ui/.env
