+++ packages/create-startupjs/create/templates/ui/.eslintrc.json
