+++ packages/create-startupjs/create/templates/ui/App.js
