+++ packages/create-startupjs/create/templates/ui/babel.config.js
