+++ packages/create-startupjs/create/templates/ui/components/TestComponent/index.js
