+++ packages/create-startupjs/create/templates/ui/config.json
