+++ packages/create-startupjs/create/templates/ui/model/TestThingModel.js
