+++ packages/create-startupjs/create/templates/ui/model/index.js
