+++ packages/create-startupjs/create/templates/ui/server/api/testThing.js
