+++ packages/create-startupjs/install/dev/package.json
