+++ packages/create-startupjs/install/init/.gitignore
