+++ packages/create-startupjs/install/init/package.json
