+++ packages/create-startupjs/install/init/startupjs.config.js
